+++ test/unit/token/Token18.test.ts
@@ -22,6 +22,22 @@ describe('Token18', () => {
     erc20 = await waffle.deployMockContract(user, IERC20Metadata__factory.abi)
   })
 
+  describe('#zero', async () => {
+    it('returns zero', async () => {
+      expect(await token18.zero()).to.equal(ethers.constants.AddressZero)
+    })
+  })
+
+  describe('#eq', async () => {
+    it('returns true', async () => {
+      expect(await token18.eq(erc20.address, erc20.address)).to.equal(true)
+    })
+
+    it('returns false', async () => {
+      expect(await token18.eq(erc20.address, ethers.constants.AddressZero)).to.equal(false)
+    })
+  })
+
   describe('#approve', async () => {
     it('approves tokens', async () => {
       await erc20.mock.allowance.withArgs(token18.address, recipient.address).returns(0)
@@ -84,12 +100,6 @@ describe('Token18', () => {
     it('returns symbol', async () => {
       await erc20.mock.symbol.withArgs().returns('TN')
       expect(await token18.connect(user).symbol(erc20.address)).to.equal('TN')
-    })
-  })
-
-  describe('#decimals', async () => {
-    it('returns decimals', async () => {
-      expect(await token18.connect(user).decimals(erc20.address)).to.equal(18)
     })
   })
